Обновил методы ОРМ, добавил проверку

+++ db/orm.py
@@ -32,8 +32,8 @@ class ORMvk:
         session = self.create_session_db()
         with session.begin():
             user = session.query(Users).filter(Users.vk_id == vk_id).first()
-
-        return user.user_id
+            if user is not None:
+                return user.user_id
 
     def add_user(self, vk_id, data):
         session = self.create_session_db()
@@ -66,14 +66,14 @@ class ORMvk:
         session.close()
         self.engine.dispose()
 
-    def add_blacklist(self, partner_vk_id):
+    def add_blacklist(self, partner_id):
         session = self.create_session_db()
         with session.begin():
-            partner = session.query(Partner).filter(Partner.partner_vk_id == partner_vk_id).first()
-            blocked = session.query(Blacklist).filter(Blacklist.partner_id == partner.partner_id).first()
+            partner = session.query(Partner).filter(Partner.partner_id == partner_id).first()
+            blocked = session.query(Blacklist).filter(Blacklist.partner_id == partner_id).first()
 
             if blocked is None:
-                new_blocked = Blacklist(user_id=partner.user_id, partner_id=partner.partner_id)
+                new_blocked = Blacklist(user_id=partner.user_id, partner_id=partner_id)
                 session.add(new_blocked)
                 session.commit()
 
@@ -81,17 +81,17 @@ class ORMvk:
         session = self.create_session_db()
         with session.begin():
             blacklist = session.query(Blacklist).filter(Blacklist.user_id == user_id).all()
+            if blacklist is not None:
+                return [item.partner_id for item in blacklist]
 
-        return [item.partner_id for item in blacklist]
-
-    def add_favorite(self, partner_vk_id):
+    def add_favorite(self, partner_id):
         session = self.create_session_db()
         with session.begin():
-            partner = session.query(Partner).filter(Partner.partner_vk_id == partner_vk_id).first()
-            favorite = session.query(Favorite).filter(Favorite.partner_id == partner.partner_id).first()
+            partner = session.query(Partner).filter(Partner.partner_id == partner_id).first()
+            favorite = session.query(Favorite).filter(Favorite.partner_id == partner_id).first()
 
             if favorite is None:
-                new_favorite = Favorite(user_id=partner.user_id, partner_id=partner.partner_id)
+                new_favorite = Favorite(user_id=partner.user_id, partner_id=partner_id)
                 session.add(new_favorite)
                 session.commit()
 
@@ -99,8 +99,8 @@ class ORMvk:
         session = self.create_session_db()
         with session.begin():
             favorite_list = session.query(Favorite).filter(Favorite.user_id == user_id).all()
-
-        return [item.partner_id for item in favorite_list]
+            if favorite_list is not None:
+                return [item.partner_id for item in favorite_list]
 
     def get_random_partner(self):
         session = self.create_session_db()
@@ -108,28 +108,31 @@ class ORMvk:
             subblack = exists().where(Blacklist.partner_id == Partner.partner_id)
             subfavor = exists().where(Favorite.partner_id == Partner.partner_id)
             query = session.query(Partner).filter(~subblack).filter(~subfavor)
-            random_row = query.offset(int(int(query.count()) * random.random())).first()
-            session.query(Users).filter(Users.user_id == random_row.user_id). \
-                update({"last_id": random_row.partner_id})
+            random_row = query.offset(int(int(query.count()) * random.random())).scalar()
+            if random_row is not None:
+                session.query(Users).filter(Users.user_id == random_row.user_id). \
+                    update({"last_id": random_row.partner_id})
 
-        return random_row.name, random_row.surname, random_row.link, random_row.foto
+                return random_row.name, random_row.surname, random_row.link, random_row.foto
 
     def get_partner(self, partner_id):
         session = self.create_session_db()
         with session.begin():
             partner = session.query(Partner).get(partner_id)
-        return partner
+            if partner is not None:
+                return partner.name, partner.surname, partner.foto, partner.link
 
     def get_search_data(self, vk_id):
         session = self.create_session_db()
-        user = session.query(Users).filter(Users.vk_id == vk_id).first()
-        if user.gender == 2:
-            gender = 1
-        else:
-            gender = 2
-        session.close()
-        self.engine.dispose()
-        return user.age, gender, user.city
+        with session.begin():
+            user = session.query(Users).filter(Users.vk_id == vk_id).first()
+            if user is not None:
+                if user.gender == 2:
+                    gender = 1
+                else:
+                    gender = 2
+
+                return user.age, gender, user.city
 
     def clear_table(self):
         session = self.create_session_db()
@@ -144,15 +147,16 @@ class ORMvk:
     def get_last_user_id(self, user_id):
         session = self.create_session_db()
         with session.begin():
-            user = session.query(Users).filter(Users.user_id == user_id).first()
-
-        return user.last_id
+            user = session.query(Users).filter(Users.user_id == user_id).scalar()
+            if user is not None:
+                return user.last_id
 
     def clear_partner_row(self, user_id):
         session = self.create_session_db()
         with session.begin():
             user = session.query(Users).filter(Users.user_id == user_id).first()
-            session.query(Partner).filter(Partner.partner_id == user.last_id).delete()
-            session.query(Users).filter(Users.user_id == user_id).update({"last_id": None})
-            session.commit()
+            if user is not None:
+                session.query(Partner).filter(Partner.partner_id == user.last_id).delete()
+                session.query(Users).filter(Users.user_id == user_id).update({"last_id": None})
+                session.commit()
